Add files via upload
Rock, Paper, Scissors is a classic hand game played between two individuals, each simultaneously forming one of three shapes with an outstretched hand. The three shapes are "rock" (a clenched fist), "paper" (an open hand), and "scissors" (a fist with the index and middle fingers extended, forming a V). The game typically consists of a series of rounds, with players selecting one of the three options in an attempt to outwit their opponent.

The rules of the game are simple:

1.Rock crushes scissors.
2.Scissors cuts paper.
3.Paper covers rock.